Move parseTags to utils
